backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
solana==0.32.0
solders==0.20.0
base58==2.1.1
pynacl==1.5.0
httpx==0.26.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Tortoise ORM (async)
tortoise-orm[asyncpg]==0.21.0
aerich==0.7.2
